requirements.txt
# Core API dependencies
fastapi==0.109.0
uvicorn==0.27.0
websockets==10.4
sqlalchemy==2.0.25
pydantic==2.5.3
python-dotenv==1.0.0
psutil==5.9.8
aiohttp==3.9.1
requests==2.31.0
httpx==0.23.3

# Data processing
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2

# Async support
asyncio==3.4.3

# External APIs
solana==0.29.2
tweepy==4.14.0
python-binance==1.0.19
